Add files via upload
• Assume that you are given a huge file consisting of billions of emails • You are asked to point some of the words in this file by providing the whole sentence that contains these words • Assume that you already found out the start and end positions of the sentences – A sorted array S[] of numbers indicating positions from the start of the text – First letter is at index 1 • Assume that you also parsed the collection and found the positions of the first letter of the words – A sorted array of W[] of numbers that contain the position

Your task is, given S[] and W[], – Find starting and ending position of each sentence that contains the word of interest – Design a quick algorithm using one of the data structures you learned. Assume size N of S[] is much greater than size M of W[]. • You should not just go through S[] in a loop to find the sentences, as it will be O(N). It will be slow for huge data even it is linear.